quantum_app_backend/requirements.txt
numpy
scipy
Flask>=2.2.3
aniso8601>=9.0.1
attrs>=22.2.0
click>=8.1.3
Flask-Cors>=3.0.10
flask-restplus>=0.13.0
flask-restx>=1.1.0
importlib-metadata>=6.1.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
jsonschema>=4.17.3
MarkupSafe>=2.1.2
pip>=23.0.1
pyrsistent>=0.19.3
pytz>=2022.7.1
setuptools>=67.6.0
six>=1.16.0
Werkzeug>=2.2.2
wheel>=0.40.0
zipp>=3.15.0
ipython>=8.17.2
odeintw
portalocker
matplotlib
flask-pymongo